Moved function to cutil folder
